Adds SpecieList components
In this commit the test uses the provider and the application store.

Through a baby steps mindset, it firstly do what is needed to turn tests
green. This commit don't change the application state when the
`store.dispatch` is called on test.

The next step is remove the duplication between the test and the
component by implementing the reducer and by connecting component
and store.