Fetches the pokemon list from the PokeAPI
The test was changed to support this. Initally the `store.dispatch`
was called directlly. Now the request is mocked.

Thinking in a behavioural way we expect that when the mount, that is,
when we shows the SpecieList on screen, we want to fetch the pokemon
list from the pokeapi and show it on the screen.

Take a look on the setupTests that is two helpers that is really
useful :)