Adds pokemon to the cart
When someone asks about where to put the tests that validates
that the cart counter starts with zero, it's easy to say "on the
cart unit test file".

But what about the tests that checks if the number changes when
an item was added to the cart?

There is a lot of ways to test it. To validate if an action was
called is easy. But it is hard to refactor, because your tests will
have a lot of stubs and mocks.

It is easy enought to create a store with some items in the cart and
validate that the number showed on the cart counter is the items'
length. But if you change your state structure, you will need to
change every test.

When we create a test, we want to validate a behaviour. To validate
when I click on "add to cart" button on our pokemon species list and
that the cart counter changes is only possible when we have both
components together. So I belive that to create an integration test
is a good approach to test that.

No store.dispatch needed, no mocks, no stubs. It is just the way that
your user interacts with our application.